refactor: split greeting responsibilities
